Don't import the kubernetes client anywhere except python-kubewatch, which is about to be gutted.
